requirements.txt
# Existing dependencies
biopython==1.84
python-pptx==0.6.23
requests==2.32.3
PyYAML==6.0.2

# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0

# Settings
pydantic-settings==2.1.0
email-validator==2.1.0

# Templates
jinja2==3.1.3

# Background tasks
apscheduler==3.10.4

# Email (production)
aiosmtplib==3.0.1

# Development & Testing
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
